Refactor rule architecture
Summary:
+ Remove Eq class
+ Add Rule class
+ Refactor substitute API
+ Replace envelope with substitute

The reason for this change is that I found it is better to handle
equations (`a = b`) just like any other function. In other words,
equations are now handled as a single expression. This makes the
code much simpler. All glue in the Eq class can be removed (in
particular the envelope stuff, which means the EqDB schema can be
simplified). The Rule class replaces Eq for usage with the
`Expr.substituteAt` method. There are now separate methods for
finding substitution positions, and substituting a rule there.
Recursive substitution is now a global function.