Drop 8bit integer storage from codec
I decided this is a case of premature optimization. It complicates
the code (including other parts since ints are stored as floating
point), and >8bit integers are acutally reasonable in the future.
If needed, it might be better to do compression using a different
method. Speed and reliability are more important here.